requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.135.2
fonttools==4.62.1
h11==0.16.0
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
MLB-StatsAPI==1.9.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
seaborn==0.13.2
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
